net/mlx4: Add user mac FW update support
Adding support for updating the FW on new port mac, when port mac change
is requested by the user. This info is required by the FW as OEM
management tools require this info directly from the NIC FW.
Check device capability bit to verify the FW supports user mac.
If the FW does support it, use set_port command to notify the FW on the
new mac.
The feature is relevant only to PF port mac.

Issue: 1093097
Change-Id: I2449db8f93f0f5270015925e8dc3c97e9656c45e
Signed-off-by: Moshe Shemesh <moshe@mellanox.com>