net/mlx5e: Unload representors in the opposite order of load
Change the cleanup order to be opposite to initialization order.

This fixes a bug in which changing mode to legacy while handling VXLAN
traffic - we hit NULL derefrence.

When unloading, for example when changing mode,
the representor encap detach routine will detach it's out-dev,i.e. the PF.
If the order of clean is not reversed - the nic detach cause the rep detach
to hit NULL deref.

Issue: 1063609
Change-Id: I609697cc684574413ad23d53cae75ea28cef9dc5
Signed-off-by: Shahar Klein <shahark@mellanox.com>