IB/mlx5: Device memory support in mlx5_ib
This patch adds the mlx5_ib provider implementation for the device
memory allocatoin API.
It implements the ib_device callbacks for allocation and deallocation
operations as well as a new mmap command support which allows mapping
an allocated device memory to a VMA.

The change also adds reporting of device memory maximum size and
alignment parameters reported in device capabilities.

The allocation/deallocation operation are making use of the mlx5_core
memic callbacks to allocated and free the actual memory on the device.

Signed-off-by: Ariel Levkovich <lariel@mellanox.com>
Signed-off-by: Leon Romanovsky <leonro@mellanox.com>