IB/uverbs: Introduce ESP steering filter and ESP_AES_GCM action
ESP is an IPsec protocol that provides both encryption and authentication.
Similarly to other IP protocols it is useful to steer traffic according
to the ESP header fields.

The ESP_AES_GCM steering action is a request for inline ESP crypto. This action
means that traffic matching the specified filter should be encrypted on ingress
or decrypted on egress by the device according to RFC 4106.

Signed-off-by: Boris Pismenny <borisp@mellanox.com>
Reviewed-by: Matan Barak <matanbo@mellanox.com>
Signed-off-by: Leon Romanovsky <leon@kernel.org>