net/mlx5: Don't store reserved part in FTEs and FGs
The current code stores fte_match_param in the software representation
of FTEs and FGs. fte_match_param contains a large reserved area at the
bottom of the struct. Since downstream patches are going to hash this
part, we would like to avoid doing so on a reserved part.

issue: 1055697
Change-Id: Id6cee97302b7ed665067037b033589acad4b1cae
Signed-off-by: Matan Barak <matanb@mellanox.com>