net/mlx5e: Check encap state when creating flow table entry
Do not create flow table entry in hardware if encap exists
but is not valid.
This may happen when encap was removed from hardware
due to neighbour state change.

Similar case when new encap is created on neighbour that is not in valid
state is handled by returning EAGAIN to upper layer.
When EAGAIN is returned, flow table entry provision to hardware is
performed by mlx5e_rep_neigh_update when neighbour becomes connected.

Issue: 1048140
Change-Id: I15ca50c94b8d3e7fc7dd6654f58ad072ce707a5a
Fixes: 232c001398ae ("net/mlx5e: Add support to neighbour update flow")
Signed-off-by: Vlad Buslov <vladbu@mellanox.com>
Reviewed-by: Roi Dayan <roid@mellanox.com>