IB/core: Add scatter end padding flags for WQ and QP
There are root complexes that are able to optimize their
performance when incoming data is multiple full cache lines.

Scatter end padding is the device's ability to pad the ending of
incoming packets (scatter) to full cache line such that the last
upstream write generated by an incoming packet will be a full cache
line.

Add a relevant entry to ib_device_cap_flags to report scatter end
padding capability of an RDMA device.

Add the QP and WQ create flags with an entry for scatter end padding:
 * A QP/WQ created with a scatter end padding flag will cause
   HW to pad the last upstream write generated by a packet to cache line.

Application logic should decide whether or not to activate this
feature according to HW details and traffic profile. E.g. if incoming
traffic is 70B and PCI cache line is 128B, it means almost x2 PCI BW.
If incoming traffic is 60B and PCI cache line is 64B, padding is not
much overhead.

Signed-off-by: Noa Osherovich <noaos@mellanox.com>
Reviewed-by: Majd Dibbiny <majd@mellanox.com>
Signed-off-by: Leon Romanovsky <leon@kernel.org>