IB/core: Introduce driver QP type
Vendors can implement type of QPs that are not described in the
InfiniBand specification. To still be able to use the IB/core layer
services (e.g. user object management) without tainting this layer with
driver proprietary logic, a new QP type is added - IB_QPT_DRIVER. This
will be a general QP type that the core layer doesn't know about its true nature.
When a command like create_qp() is passed to a hardware driver the extra
data that is required is taken from the driver channel.
Downstream patches from this series will use that QP type in the mlx5
driver.

issue: 1228326
Change-Id: I6660be1b3cc2e38963aedb81e805f3b06b536182
Signed-off-by: Moni Shoua <monis@mellanox.com>