IB/core: Refactor allocation code to increase number of char device
There is a need to increase number of possible char devices to support
large number of SR-IOV instances. The current limit is in range of
64-128 devices/ports and this refactoring patch increases it to support
upto 1024.

The patch performs the following steps to refactor the code:
1. Removes the split bitmap for fixed and overflow dev numbers.
2. Pre-allocates the non-legacy major number range during driver
initialization for the simplicity.
3. Add new define (RDMA_MAX_PORTS) that is shared between all drivers
for the future expansion of non-legacy minor number range. This is the
maximum total number of ports on all struct ib_devices.
4. Set RDMA_MAX_PORTS to 1024.

Signed-off-by: Huy Nguyen <huyn@mellanox.com>
Signed-off-by: Leon Romanovsky <leon@kernel.org>