IB/core: Honor port_num while resolving GID for IB link layer
Currently initialized ah_attr contains the port number to which cm_id is
bound. However, while searching for GID table for matching GID entry,
port number is ignored.
It is better to honor the port number of the cm_id while searching GID
table for IB link layer as well.

Change-Id: Ie11d508de87e9a565e10ca3bcc3e13c0352cfde1
Reviewed-by: Daniel Jurgens <danielj@mellanox.com>
Signed-off-by: Parav Pandit <parav@mellanox.com>
Signed-off-by: Leon Romanovsky <leon@kernel.org>